--- src/main/java/com/example/nexufy/service/UserService.java
@@ -1,38 +1,0 @@
-package com.example.nexufy.service;
-
-import com.example.nexufy.persistence.entities.User;
-import com.example.nexufy.persistence.repository.UserRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.bson.types.ObjectId;
-
-import java.util.List;
-import java.util.Optional;
-
-@Service
-public class UserService {
-
-    @Autowired
-    private UserRepository userRepository;
-
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
-    }
-
-    public Optional<User> getUserById(String id) {
-        return userRepository.findById(id);
-    }
-
-    public User addUser(User user) {
-        return userRepository.save(user);
-    }
-
-    public void deleteUser(String id) {
-        userRepository.deleteById(id);
-    }
-
-    public User updateUser(String id, User user) {
-        user.setId(String.valueOf(new ObjectId(id)));
-        return userRepository.save(user);
-    }
-}
