+++ src/main/resources/application.properties
@@ -1,8 +1,4 @@
 spring.application.name=nexufy
-spring.h2.console.enabled=true
-spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.url=jdbc:h2:tcp://localhost/~/test
-spring.datasource.username=sa
-spring.datasource.password=
-spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=create
+spring.data.mongodb.database=Nexufy
+spring.data.mongodb.uri=mongodb+srv://gaspartorres352:Sqql8l7aFIG38ymQ@nexufy.3yde8dt.mongodb.net/
+
